Separate items collection by User

+++ lib/repositories/items.repository.dart
@@ -1,10 +1,18 @@
 import 'package:cantwait28/models/item_model.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 
 class ItemsRepository {
   Stream<List<ItemModel>> getItemsStream() {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+
+    if (userID == null) {
+      throw Exception('User is not logged in');
+    }
     return FirebaseFirestore.instance
-        .collection('items')
+        .collection('items1')
+        .doc(userID)
+        .collection('items1')
         .orderBy('release_date')
         .snapshots()
         .map(
@@ -23,37 +31,60 @@ class ItemsRepository {
     );
   }
 
-  Future<void> delete({required String id})  {
-  return FirebaseFirestore.instance
-          .collection('items')
-          .doc(id)
-          .delete();
+  Future<void> delete({required String id}) {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+
+    if (userID == null) {
+      throw Exception('User is not logged in');
+    }
+    return FirebaseFirestore.instance
+        .collection('items1')
+        .doc(userID)
+        .collection('items1')
+        .doc(id)
+        .delete();
   }
 
-   Future<ItemModel> get({required String id}) async {
-  final doc = await FirebaseFirestore.instance
-          .collection('items')
-          .doc(id)
-          .get();
-          return ItemModel(
-              id: doc.id,
-              title: doc['title'],
-              imageURL: doc['image_url'],
-              releasedate: (doc['release_date'] as Timestamp).toDate(),
-            );
+  Future<ItemModel> get({required String id}) async {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+
+    if (userID == null) {
+      throw Exception('User is not logged in');
+    }
+    final doc = await FirebaseFirestore.instance
+        .collection('items1')
+        .doc(userID)
+        .collection('items1')
+        .doc(id)
+        .get();
+    return ItemModel(
+      id: doc.id,
+      title: doc['title'],
+      imageURL: doc['image_url'],
+      releasedate: (doc['release_date'] as Timestamp).toDate(),
+    );
   }
 
- Future<void> add(
+  Future<void> add(
     String title,
     String imageURL,
     DateTime releaseDate,
   ) async {
-      await FirebaseFirestore.instance.collection('items').add(
-        {
-          'title': title,
-          'image_url': imageURL,
-          'release_date': releaseDate,
-        },
-      );
-}
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+
+    if (userID == null) {
+      throw Exception('User is not logged in');
+    }
+    await FirebaseFirestore.instance
+        .collection('items1')
+        .doc(userID)
+        .collection('items1')
+        .add(
+      {
+        'title': title,
+        'image_url': imageURL,
+        'release_date': releaseDate,
+      },
+    );
+  }
 }
